move DavitConstants.ts in src directory
